+++ yandex/data_source_yandex_compute_instance_group.go
@@ -702,6 +702,11 @@ func dataSourceYandexComputeInstanceGroup() *schema.Resource {
 				Type:     schema.TypeString,
 				Computed: true,
 			},
+
+			"deletion_protection": {
+				Type:     schema.TypeBool,
+				Computed: true,
+			},
 		},
 	}
 }